# openapi/licoreria.yaml
openapi: 3.1.1
info:
  title: API Licorería / Discoteca
  description: >-
    Contrato generado desde el código de la API del sistema de licorería y discoteca (sucursal única). Los errores se
    devuelven conforme a RFC 7807.
  version: v1
paths:
...
  /api/v1/mesas:
    get:
      tags:
        - Mesas
      parameters:
        - name: zonaId
          in: query
          schema:
            type: string
            format: uuid
        - name: fechaHora
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MesaDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MesaDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MesaDto'
    post:
      tags:
        - Mesas
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MesaCrearDto'
          text/json:
            schema:
              $ref: '#/components/schemas/MesaCrearDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MesaCrearDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MesaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/MesaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/MesaDto'
  /api/v1/mesas/{id}:
    put:
      tags:
        - Mesas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MesaEditarDto'
          text/json:
            schema:
              $ref: '#/components/schemas/MesaEditarDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MesaEditarDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MesaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/MesaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/MesaDto'
    delete:
      tags:
        - Mesas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /api/v1/mesas/{id}/desalojar:
    post:
      tags:
        - Mesas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/productos:
    get:
      tags:
        - Productos
      parameters:
        - name: Page
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: PageSize
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: PaginaNormalizada
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: TamanoNormalizado
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: categoriaId
          in: query
          schema:
            type: string
            format: uuid
        - name: busqueda
          in: query
          schema:
            type: string
        - name: activo
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ResultadoPaginadoOfProductoDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ResultadoPaginadoOfProductoDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ResultadoPaginadoOfProductoDto'
    post:
      tags:
        - Productos
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoCrearDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ProductoCrearDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProductoCrearDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProductoDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProductoDto'
  /api/v1/productos/{id}:
    get:
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProductoDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProductoDto'
    put:
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoEditarDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ProductoEditarDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProductoEditarDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProductoDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProductoDto'
    delete:
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /api/v1/productos/{id}/modificadores:
    get:
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoModificadorDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoModificadorDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoModificadorDto'
    post:
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarModificadorDto'
          text/json:
            schema:
              $ref: '#/components/schemas/AsignarModificadorDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AsignarModificadorDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProductoModificadorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoModificadorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProductoModificadorDto'
  /api/v1/productos/{id}/modificadores/{productoModificadorId}:
    delete:
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: productoModificadorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/sesiones-caja:
    post:
      tags:
        - SesionesCaja
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbrirCajaDto'
          text/json:
            schema:
              $ref: '#/components/schemas/AbrirCajaDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AbrirCajaDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
    get:
      tags:
        - SesionesCaja
      parameters:
        - name: Page
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: PageSize
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: PaginaNormalizada
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: TamanoNormalizado
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ResultadoPaginadoOfSesionCajaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ResultadoPaginadoOfSesionCajaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ResultadoPaginadoOfSesionCajaDto'
...
  /api/v1/sesiones-caja/{id}/movimientos:
    post:
      tags:
        - SesionesCaja
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimientoCajaCrearDto'
          text/json:
            schema:
              $ref: '#/components/schemas/MovimientoCajaCrearDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MovimientoCajaCrearDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
  /api/v1/sesiones-caja/{id}/cerrar:
    post:
      tags:
        - SesionesCaja
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CerrarCajaDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CerrarCajaDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CerrarCajaDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/SesionCajaDto'
...
components:
  schemas:
...
    AbrirCajaDto:
      required:
        - fondoInicial
      type: object
      properties:
        fondoInicial:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
...
    AreaDestino:
      type: integer
...
    AsignarModificadorDto:
      required:
        - modificadorId
      type: object
      properties:
        modificadorId:
          type: string
          format: uuid
        minimo:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
          default: 0
        maximo:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
          default: 1
        requerido:
          type: boolean
          default: false
...
    EstadoSesionCaja:
      type: integer
...
    MesaCrearDto:
      required:
        - zonaId
        - numero
        - capacidad
      type: object
      properties:
        zonaId:
          type: string
          format: uuid
        numero:
          type: string
        capacidad:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        forma:
          type: string
          default: redonda
        posX:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
        posY:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
        ancho:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 1
        alto:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 1
    MesaDto:
      required:
        - id
        - zonaId
        - zonaNombre
        - numero
        - capacidad
        - forma
        - posX
        - posY
        - ancho
        - alto
        - activa
        - disponible
      type: object
      properties:
        id:
          type: string
          format: uuid
        zonaId:
          type: string
          format: uuid
        zonaNombre:
          type: string
        numero:
          type: string
        capacidad:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        forma:
          type: string
        posX:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        posY:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        ancho:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        alto:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        activa:
          type: boolean
        disponible:
          type: boolean
        cuentaId:
          type:
            - 'null'
            - string
          format: uuid
        reservada:
          type: boolean
          default: false
    MesaEditarDto:
      required:
        - id
        - zonaId
        - numero
        - capacidad
        - forma
        - posX
        - posY
        - ancho
        - alto
        - activa
      type: object
      properties:
        id:
          type: string
          format: uuid
        zonaId:
          type: string
          format: uuid
        numero:
          type: string
        capacidad:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        forma:
          type: string
        posX:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        posY:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        ancho:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        alto:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        activa:
          type: boolean
...
    ProductoCrearDto:
      required:
        - nombre
        - descripcion
        - categoriaId
        - marcaId
        - impuestoId
        - tipo
        - areaDestino
        - gradoAlcoholico
        - imagenUrl
        - variantes
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type:
            - 'null'
            - string
        categoriaId:
          type: string
          format: uuid
        marcaId:
          type:
            - 'null'
            - string
          format: uuid
        impuestoId:
          type:
            - 'null'
            - string
          format: uuid
        tipo:
          $ref: '#/components/schemas/TipoProducto'
        areaDestino:
          $ref: '#/components/schemas/AreaDestino'
        gradoAlcoholico:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - 'null'
            - number
            - string
          format: double
        imagenUrl:
          type:
            - 'null'
            - string
        variantes:
          type: array
          items:
            $ref: '#/components/schemas/VarianteCrearDto'
    ProductoDto:
      required:
        - id
        - nombre
        - descripcion
        - categoriaId
        - categoriaNombre
        - marcaId
        - marcaNombre
        - impuestoId
        - tipo
        - areaDestino
        - gradoAlcoholico
        - imagenUrl
        - activo
        - variantes
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        descripcion:
          type:
            - 'null'
            - string
        categoriaId:
          type: string
          format: uuid
        categoriaNombre:
          type: string
        marcaId:
          type:
            - 'null'
            - string
          format: uuid
        marcaNombre:
          type:
            - 'null'
            - string
        impuestoId:
          type:
            - 'null'
            - string
          format: uuid
        tipo:
          $ref: '#/components/schemas/TipoProducto'
        areaDestino:
          $ref: '#/components/schemas/AreaDestino'
        gradoAlcoholico:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - 'null'
            - number
            - string
          format: double
        imagenUrl:
          type:
            - 'null'
            - string
        activo:
          type: boolean
        variantes:
          type: array
          items:
            $ref: '#/components/schemas/ProductoVarianteDto'
    ProductoEditarDto:
      required:
        - id
        - nombre
        - descripcion
        - categoriaId
        - marcaId
        - impuestoId
        - tipo
        - areaDestino
        - gradoAlcoholico
        - imagenUrl
        - activo
        - variantes
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        descripcion:
          type:
            - 'null'
            - string
        categoriaId:
          type: string
          format: uuid
        marcaId:
          type:
            - 'null'
            - string
          format: uuid
        impuestoId:
          type:
            - 'null'
            - string
          format: uuid
        tipo:
          $ref: '#/components/schemas/TipoProducto'
        areaDestino:
          $ref: '#/components/schemas/AreaDestino'
        gradoAlcoholico:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - 'null'
            - number
            - string
          format: double
        imagenUrl:
          type:
            - 'null'
            - string
        activo:
          type: boolean
        variantes:
          type: array
          items:
            $ref: '#/components/schemas/VarianteCrearDto'
    ProductoModificadorDto:
      required:
        - id
        - modificadorId
        - modificadorNombre
        - precioAdicional
        - minimo
        - maximo
        - requerido
      type: object
      properties:
        id:
          type: string
          format: uuid
        modificadorId:
          type: string
          format: uuid
        modificadorNombre:
          type: string
        precioAdicional:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        minimo:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        maximo:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        requerido:
          type: boolean
...
    ResultadoPaginadoOfProductoDto:
      required:
        - items
        - page
        - pageSize
        - totalItems
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductoDto'
        page:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        totalItems:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
...
    ResultadoPaginadoOfSesionCajaDto:
      required:
        - items
        - page
        - pageSize
        - totalItems
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SesionCajaDto'
        page:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        totalItems:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
...
    SesionCajaDto:
      required:
        - id
        - estado
        - fondoInicial
        - montoEsperado
        - montoContado
        - descuadre
        - ventasDelTurnoUSD
        - cuentasDesalojadas
        - abiertaEn
        - cerradaEn
        - movimientos
        - arqueo
      type: object
      properties:
        id:
          type: string
          format: uuid
        estado:
          $ref: '#/components/schemas/EstadoSesionCaja'
        fondoInicial:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        montoEsperado:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        montoContado:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        descuadre:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        ventasDelTurnoUSD:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        cuentasDesalojadas:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        abiertaEn:
          type: string
          format: date-time
        cerradaEn:
          type:
            - 'null'
            - string
          format: date-time
        movimientos:
          type: array
          items:
            $ref: '#/components/schemas/MovimientoCajaDto'
        arqueo:
          type: array
          items:
            $ref: '#/components/schemas/ArqueoLineaDto'
...
    TipoProducto:
      type: integer
...
  securitySchemes:
    Bearer:
      type: http
      description: 'Ingrese el token JWT. Ejemplo: Bearer {token}'
      scheme: bearer
      bearerFormat: JWT
security:
  - {}
...